Added Agent.py with Agent Class and constants, which each algo script imports. Separated out functions from APF
